feat(PiecesDto): 기물 dto 리스트 구현
