+++ src/main/java/chess/web/controller/BoardController.java
@@ -1,18 +1,13 @@
 package chess.web.controller;
 
 import chess.domain.board.Board;
-import chess.domain.board.File;
 import chess.domain.board.Position;
-import chess.domain.board.Rank;
 import chess.domain.piece.vo.TeamColor;
 import chess.web.dto.PieceDto;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Comparator;
+import chess.web.dto.PiecesDto;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 import spark.ModelAndView;
 import spark.Request;
 import spark.Response;
@@ -28,36 +23,12 @@ public class BoardController {
     }
 
     public Object printCurrentBoard(Request request, Response response) {
-        List<PieceDto> pieces = new ArrayList<>();
-        for (Position position : getAllPosition()) {
-            addPieceDto(pieces, position);
-        }
-        return render(toModel(pieces), "board.html");
-    }
-
-    private List<Position> getAllPosition() {
-        List<Position> positions = new ArrayList<>();
-        for (Rank rank : getReverseOrderRanks()) {
-            positions.addAll(Arrays.stream(File.values())
-                    .map(file -> Position.of(file, rank))
-                    .collect(Collectors.toList()));
-        }
-        return positions;
-    }
-
-    private void addPieceDto(List<PieceDto> pieces, Position position) {
-        try {
-            pieces.add(PieceDto.of(board.findPieceInPosition(position)));
-        } catch (IllegalArgumentException e) {
-            pieces.add(new PieceDto("", ""));
-        }
-    }
-
-    private Map<String, Object> toModel(List<PieceDto> pieces) {
         Map<String, Object> model = new HashMap<>();
+        List<PieceDto> pieces = PiecesDto.of(board)
+                .getValue();
         model.put("pieces", pieces);
         model.put("turnColor", board.getCurrentTurnTeamColor());
-        return model;
+        return render(model, "board.html");
     }
 
     public String move(Request request, Response response) {
@@ -83,11 +54,5 @@ public class BoardController {
 
     private static String render(Map<String, Object> model, String templatePath) {
         return new HandlebarsTemplateEngine().render(new ModelAndView(model, templatePath));
-    }
-
-    private static List<Rank> getReverseOrderRanks() {
-        return Arrays.stream(Rank.values())
-                .sorted(Comparator.reverseOrder())
-                .collect(Collectors.toList());
     }
 }
